+++ packages/brick_rest_generators/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Unreleased
 
+* Fix deserialize and serialize `enumAsString` for non-nullable types (#183)
+
 ## 1.0.0+3
 
 * Fix pubspec dependencies